Make retention/deletion optional

+++ entrypoint.sh
@@ -24,15 +24,17 @@ echo "Starting backup"
 DATE=$(date +%Y.%m.%d-%H.%M.%S)
 curl -s -X PUT "${CLUSTER_URL}/_snapshot/${REPOSITORY_NAME}/${DATE}?wait_for_completion=true&pretty"
 
-echo
-echo "Finding snapshots older than the retention threshold (${SNAPSHOT_RETENTION})"
-SNAPSHOTS_TO_DELETE=$(curl -s ${CLUSTER_URL}/_snapshot/${REPOSITORY_NAME}/_all | jq -r ".snapshots[:-${SNAPSHOT_RETENTION}][].snapshot")
-
-if [[ ! -z "${SNAPSHOTS_TO_DELETE}" ]]; then
-    for snapshot in ${SNAPSHOTS_TO_DELETE}; do
-        echo "Deleting snapshot ${snapshot}"
-        curl -s -X DELETE ${CLUSTER_URL}/_snapshot/${REPOSITORY_NAME}/${snapshot}?pretty
-    done
+if [[ ! -z "${SNAPSHOT_RETENTION}" ]]; then
+    echo
+    echo "Finding snapshots older than the retention threshold (${SNAPSHOT_RETENTION})"
+    SNAPSHOTS_TO_DELETE=$(curl -s ${CLUSTER_URL}/_snapshot/${REPOSITORY_NAME}/_all | jq -r ".snapshots[:-${SNAPSHOT_RETENTION}][].snapshot")
+    
+    if [[ ! -z "${SNAPSHOTS_TO_DELETE}" ]]; then
+        for snapshot in ${SNAPSHOTS_TO_DELETE}; do
+            echo "Deleting snapshot ${snapshot}"
+            curl -s -X DELETE ${CLUSTER_URL}/_snapshot/${REPOSITORY_NAME}/${snapshot}?pretty
+        done
+    fi
 fi
 
 echo
